// src/app/models/banner.model.ts
// Generated by https://quicktype.io

export interface Banner {
  idBanner:         number;
  idProveedor:      number;
  contenidoAnuncio: string;
  pathImagen:       string;
  fileBytes:        null;
  linkAnuncio:      string;
  rangoFechas:      string;
  orden:            number;
  prioridad:        number;
  enviarNoCliente:  boolean;
  enviarCliente:    boolean;
  plataforma:       string;
  fechaCreacion:    string;
  fechaCreacionUtc: string;
  idProveedorWeb:   number;
  idUsuarioWeb:     number;
  proveedorWeb:     null;
}